requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.2.0
numpy>=1.26.2,<2.0.0
scikit-learn==1.4.0
psycopg2-binary==2.9.9
paramiko==3.4.0
python-multipart==0.0.6
catboost==1.2.7
lightgbm==4.6.0
optuna==4.4.0
sqlalchemy==2.0.30
requests==2.32.2
python-dotenv==0.21.0              